Re-render counts on page visibility change
This is useful to update the current day when loading the app, and
avoids using an interval to re-render the page periodically, which would
be very wasteful since it really only changes at 5am every morning.

To do this, the Vue app was wrapped in an App.vue component, which holds
a current date, and refreshes it on the "visibilitychange" window event.

All functions using a `new Date()` now use the current date passed as a
prop (it will make them easier to test anyway).